+++ client/packages/common/src/ui/forms/JsonForms/useJsonForms.tsx
@@ -1,9 +1,5 @@
 import React, { useEffect, useState } from 'react';
-import {
-  Typography,
-  UnhappyMan,
-  useNavigate,
-} from '@openmsupply-client/common';
+import { RouteBuilder, useNavigate } from '@openmsupply-client/common';
 import {
   Box,
   DialogButton,
@@ -12,80 +8,13 @@ import {
   useTranslation,
   useNotification,
   useConfirmOnLeaving,
-  BasicSpinner,
 } from '@openmsupply-client/common';
-import { JsonForms } from '@jsonforms/react';
-import {
-  JsonFormsRendererRegistryEntry,
-  JsonSchema,
-  UISchemaElement,
-} from '@jsonforms/core';
-import { materialRenderers } from '@jsonforms/material-renderers';
-import {
-  stringTester,
-  TextField,
-  selectTester,
-  Selector,
-  groupTester,
-  Group,
-  labelTester,
-  Label,
-  dateTester,
-  Date,
-  arrayTester,
-  Array,
-} from './components';
+import { JsonSchema, UISchemaElement } from '@jsonforms/core';
+
 import { useDocument } from './api';
 import { DocumentRegistryFragment } from './api/operations.generated';
-
-export type JsonData = {
-  [key: string]: string | number | boolean | null | unknown | JsonData;
-};
-
-interface JsonFormsComponentProps {
-  data: JsonData;
-  jsonSchema: JsonSchema;
-  uiSchema: UISchemaElement;
-  setData: (data: JsonData) => void;
-  setError: (error: string | false) => void;
-  renderers: JsonFormsRendererRegistryEntry[];
-}
-
-const FormComponent = ({
-  data,
-  jsonSchema,
-  uiSchema,
-  setData,
-  setError,
-  renderers,
-}: JsonFormsComponentProps) => {
-  return (
-    <JsonForms
-      schema={jsonSchema}
-      uischema={uiSchema}
-      data={data}
-      renderers={renderers}
-      // cells={materialCells}
-      onChange={({ errors, data }) => {
-        setData(data);
-        if (errors && errors.length) {
-          setError(errors?.map(({ message }) => message ?? '').join(', '));
-          console.warn('Errors: ', errors);
-        } else {
-          setError(false);
-        }
-      }}
-    />
-  );
-};
-
-// Prevents Form window being loaded with the same scroll position as its parent
-const ScrollFix = () => {
-  useEffect(() => {
-    document.getElementById('document-display')?.scrollIntoView();
-  }, []);
-  return null;
-};
+import { JsonData, JsonForm } from './JsonForm';
+import { AppRoute } from '@openmsupply-client/config';
 
 export type SavedDocument = {
   id: string;
@@ -112,7 +41,7 @@ interface JsonFormOptions {
  * Information required to create a new document
  */
 export interface CreateDocument {
-  data: any;
+  data: JsonData;
   documentRegistry: DocumentRegistryFragment;
 }
 
@@ -126,7 +55,6 @@ export const useJsonForms = (
   const [documentId, setDocumentId] = useState<string | undefined>();
   // document name can change from the input parameter when creating a new document
   const [documentName, setDocumentName] = useState<string | undefined>(docName);
-
   const [documentRegistry, setDocumentRegistry] = useState<{
     formSchemaId: string;
     jsonSchema: JsonSchema;
@@ -150,6 +78,82 @@ export const useJsonForms = (
     isLoading,
     isError,
   } = useDocument.get.document(documentName ?? '', !!documentName);
+
+  const {
+    showButtonPanel = true,
+    onCancel = () => navigate(RouteBuilder.create(AppRoute.Patients).build()),
+    saveConfirmationMessage = t('messages.confirm-save-generic'),
+    cancelConfirmationMessage = t('messages.confirm-cancel-generic'),
+    saveSuccessMessage = t('success.data-saved'),
+  } = options;
+
+  useConfirmOnLeaving(isDirty);
+
+  const saveData = async () => {
+    if (data === undefined) {
+      return;
+    }
+    setSaving(true);
+
+    // Run mutation...
+    try {
+      const result = await options.handleSave?.(
+        data,
+        documentRegistry.formSchemaId,
+        documentId
+      );
+
+      setDocumentName(result?.name);
+      setIsDirty(false);
+
+      const successSnack = success(saveSuccessMessage);
+      successSnack();
+    } catch (err) {
+      const errorSnack = errorNotification(t('error.problem-saving'));
+      errorSnack();
+    } finally {
+      setSaving(false);
+    }
+  };
+
+  const updateData = (newData: JsonData) => {
+    setIsDirty(isDirty === undefined ? false : true);
+    setData(newData);
+  };
+
+  const showSaveConfirmation = useConfirmationModal({
+    onConfirm: saveData,
+    message: saveConfirmationMessage,
+    title: t('heading.are-you-sure'),
+  });
+
+  const showCancelConfirmation = useConfirmationModal({
+    onConfirm: onCancel,
+    message: cancelConfirmationMessage,
+    title: t('heading.are-you-sure'),
+  });
+
+  const ButtonPanel = () => (
+    <Box id="button-panel" paddingBottom={5} display="flex" gap={5}>
+      <LoadingButton
+        onClick={() => showSaveConfirmation()}
+        isLoading={saving}
+        disabled={error !== false || isLoading || !isDirty}
+        color="secondary"
+      >
+        {t('button.save')}
+      </LoadingButton>
+      <DialogButton
+        variant="cancel"
+        disabled={isLoading}
+        onClick={() => {
+          if (isDirty) showCancelConfirmation();
+          else onCancel();
+        }}
+      />
+    </Box>
+  );
+
   useEffect(() => {
     if (!databaseResponse) return;
 
@@ -181,141 +185,19 @@ export const useJsonForms = (
     }
   }, [createDoc]);
 
-  const {
-    showButtonPanel = true,
-    onCancel = () => navigate(-1),
-    saveConfirmationMessage = t('messages.confirm-save-generic'),
-    cancelConfirmationMessage = t('messages.confirm-cancel-generic'),
-    saveSuccessMessage = t('success.data-saved'),
-  } = options;
-
-  useConfirmOnLeaving(isDirty);
-
-  const updateData = (newData: JsonData) => {
-    setIsDirty(isDirty === undefined ? false : true);
-    setData(newData);
-  };
-
-  const saveData = async () => {
-    if (data === undefined) {
-      return;
-    }
-    setSaving(true);
-    console.log('Saving data...');
-
-    // Run mutation...
-    try {
-      const result = await options.handleSave?.(
-        data,
-        documentRegistry.formSchemaId,
-        documentId
-      );
-
-      setDocumentName(result?.name);
-      setIsDirty(false);
-
-      const successSnack = success(saveSuccessMessage);
-      successSnack();
-    } catch (err) {
-      const errorSnack = errorNotification(t('error.problem-saving'));
-      errorSnack();
-    } finally {
-      setSaving(false);
-    }
-  };
-
-  const renderers = [
-    // We should be able to remove materialRenderers once we are sure we have custom components to cover all cases.
-    ...materialRenderers,
-    { tester: stringTester, renderer: TextField },
-    { tester: selectTester, renderer: Selector },
-    { tester: groupTester, renderer: Group },
-    { tester: labelTester, renderer: Label },
-    { tester: dateTester, renderer: Date },
-    { tester: arrayTester, renderer: Array },
-  ];
-
-  const showSaveConfirmation = useConfirmationModal({
-    onConfirm: saveData,
-    message: saveConfirmationMessage,
-    title: t('heading.are-you-sure'),
-  });
-
-  const showCancelConfirmation = useConfirmationModal({
-    onConfirm: onCancel,
-    message: cancelConfirmationMessage,
-    title: t('heading.are-you-sure'),
-  });
-
-  const ButtonPanel = () => (
-    <Box id="button-panel" paddingBottom={5} display="flex" gap={5}>
-      <LoadingButton
-        onClick={() => {
-          if (isDirty) showSaveConfirmation();
-          else onCancel();
-        }}
-        isLoading={saving}
-        disabled={error !== false || isLoading}
-        color="secondary"
-      >
-        {t('button.save')}
-      </LoadingButton>
-      <DialogButton
-        variant="cancel"
-        disabled={isLoading}
-        onClick={() => {
-          if (isDirty) showCancelConfirmation();
-          else onCancel();
-        }}
-      />
-    </Box>
-  );
-
-  const JsonForm =
-    //  isLoading || !!data ? (
-    isError ? (
-      <Box
-        display="flex"
-        flexDirection="column"
-        justifyContent="center"
-        alignItems="center"
-        width="100%"
-        gap={2}
-      >
-        <UnhappyMan />
-        <Typography color="error">{t('error.unable-to-load-data')}</Typography>
-      </Box>
-    ) : (
-      <Box
-        id="document-display"
-        display="flex"
-        flexDirection="column"
-        justifyContent={!data ? 'flex-end' : 'flex-start'}
-        alignItems="center"
-        width="100%"
-        gap={2}
-        paddingX={10}
-      >
-        <ScrollFix />
-        {isLoading || !data ? (
-          <BasicSpinner />
-        ) : (
-          <FormComponent
-            data={data}
-            jsonSchema={documentRegistry.jsonSchema}
-            uiSchema={documentRegistry.uiSchema}
-            setData={updateData}
-            setError={setError}
-            renderers={renderers}
-          />
-        )}
-        {showButtonPanel && <ButtonPanel />}
-      </Box>
-      // ) : (
-    );
-
   return {
-    JsonForm,
+    JsonForm: (
+      <JsonForm
+        data={data}
+        documentRegistry={documentRegistry}
+        isError={isError}
+        isLoading={isLoading}
+        setError={setError}
+        updateData={updateData}
+      >
+        {showButtonPanel && <ButtonPanel />}
+      </JsonForm>
+    ),
     saveData,
     loading: isLoading,
     error,
